refs #11347 Advanced ScheduleRealTimeUpdateAction
In progress development to create an advanced version of ScheduleRealTimeUpdateAction
Once complete, it will either be merger in ScheduleRealTimeUpdateAction or renamed to
a relevant name.
This action reschedules a service based on a given Arrival or Departure time.
If the next services end up being before the current one, they are rescheduled as
well.
Issue : #11347