bug refs #27002 added basic support for xulrunner
Our kioskPlayer is not able to configure the proxy for custom browsers
This patch add a new browser category to choose from in the kiosk admin
site on port 5000. It does nothing more than the custom browser than adding
a message in the logs to explain how to configure a xulrunner browser
for the proxy.