fixes bug refs #27002 improved kiosk to include a local fallback cache
This version no more use kiosk.html to render the site. The python part
does the job to check if SYNTHESE is accessible. If not it changes
the browser URL to display the fallback site from a local cache.

The python part itself is used to serve the pages in HTTP.

The wiki has been updated to reflect the new operational mode:
https://extranet.rcsmobility.com/projects/synthese/wiki/Synthese_Kiosk