added support for an auto refresh mode in the kiosk
Now the kiosk can refresh periodically the browser. To activate
this mode you must set a time in seconds in the key 'auto_refresh_timeout'
in the config.json file.

It works best if you also install a selenium webdriver for the
browser you are using. In this case a refresh is done without
restarting the browser.

bug refs #27002