fixed getNextService for services after midnight
This new patch fixes the issue by filling properly the
by hour time slots in _departureIndex and _arrivalIndex.

Before this patch, a service at 24h30 would be saved in
_departureIndex[0]. Now it may be saved in _departureIndex[23]
if there is no services in the 23 hour range. If there is a service
at 23 hour, then it is not a problem because getNextServices will
find it because it iterates over the service list until the end
of the list.