Added support for disabling theoretical time
By default this patch does not change anything.

Now it is possible to set the params enable_real_time and
enable_theoretical on the command line (admin missing).

When the theoretical time is disabled, if the real_time is
enable then we will take in account only services having
real time data.

This feature is used only in PTTimeSlotRoutePlanner.

bug refs #25907