improved crash sequence.
Now we stop our HTTP Server before stopping all the services to
be sure no request we get in while shutting down.

And we stop the services in reverse order so that 10_DB for example
is stopped the last as other services may potentially still use
it.

There is a bug because in ascii order the 102_svn module is before
10_db.

I still have crash in GetDB while shutting down.