Fixed calculation of the trip duration when a scheduled service is Real Time.
We were not calculating the trip duration properly when a
scheduled service had real time information. In this case, the
trip duration was calculated from the theoretical start instead
of the real time start. 

Thus for example a trip starting at 10:00 theoretical with a 15
minutes duration, when shifted to 10h15 Real Time was reporting a
trip duration of 30 minutes.

The ServicePointer getTheoreticalArrivalDateTime() was also wrong.

Added a test for this in ServiceTest.