Use of ParameterMap in HTMLTable and fix dates check for commercial line admin
When using HTML Table, the GET parameters were not used by the setParameterMap of ResultHTMLTable because they are in the Fields attribute and only the <string, string> map was given.

In commercial line admin, the dates check did not work when the line used external calendars because of the link level. The link level is now explicitly given when searching the scheduled and continuous services. Fixes bug refs #28381