Service tdg: added support for passing the generation_method
Added the parameter 'generation_method' to specify the generation method.
Must be one of 'standard_method', 'with_forced_destinations_method',
'route_planning' or 'display_children_only'. The default is 'standard_method'

We miss some option to use the full potential of each generation method.

This is backward compatible.

Added inline documentation of the TDG service based on the data
found on the wiki.