Revision r8836 merged from branch 3.8.beta1
kiosk, replaced online check by a valid call to SYNTHESE

This valid call is also checked for an OK HTTP status.
If not OK we go in offline mode.

Added an option to let the javascript call us to force the
offline mode and request an url to check. This way, if
the javascript fails to load the url, it calls the python
part requiring to go offline and it passes the url to check.

bug refs #27002