Revision r8865 merged from branch 3.8.beta1
fixed site svn export

For some reasons it happens that site page may have the same rank.
In this case these pages was lost in the svn commit process.

this is due to our TreeNode storing items in a map thus hidding
items at the same tree rank.

Since the TreeNode is used in other places, there is a risk of
regression in DisplayScreen and CalendarTemplate.