changed http server to no more create threads on demand.
Asio does not require us to do this. It will pick the next working
thread as soon as one becomes ready. Adding more threads, especially
more than the number of CPU core won't improve performance in our case
because in most cases the workload is CPU bound.

When running SYNTHESE on a server, it must be configured with
--param nb_threads=<my number of cpu core>

By default this number is 5.