Fixed build svn revision and date.
The build svn revision and date was generated once at cmake
time. After an svn update the svn revision was not recalculated
and thus the version displayed at runtime was misleading.

This patch calculate the version at make time. In order to not
have SYNTHESE being rebuild at each make, the new generated
svnversion.h is only updated if the svn version as changed. The
build date is now the last svn modification date which is not
perfect but better than the cmake date.