Merged Revision r9040 from 3.8.beta1
svn site save, fixed end of line to be always unix style.

In .dump and .name file we were using the default end of line
style of the runnning SYNTHESE. Thus a svn site created on
windows had the wrong end of line. This was creating confusing
changes while saving a site on a platform or another.