Bug Fix on two fields autocompleter.
If a StopArea and a Road Place had exactly the same name (case sensitive),
it would be added only one time into allPlacesMatcher of the city,
since it's represented by a map, with the full name for key.
Now we are searching into StopArea matcher then into RoadPlace matcher and sorting the results afterward.