Fixed IneoRT, now the path is properly saved.
There was a bug in which the path was not loaded in
FIELDS_ONLY_LOAD_LEVEL mode and thus was saved as 0.

Added a hack by using the linkLevel UNKNOWN_LOAD_LEVEL
in JourneyPatternTableSync to let it don't load
anythink. One can think of it as create the object with
its key only.