BugFix : in RoadPlace::getHouse, revert commit r4833
If the house number isn't in any bounds we simply return a house on the chunk 
with the closest house number bounds, and not the next one sometimes.
This doesn't had much sense and it was leading to Adresses on chunk without any geometry,
causing them to have no geometry either.