Feature : new parameters on StopPointsListFunction.
SYNTHESE was looking for a service on a stop point in the current day,
we can choose how much days SYNTHESE checks for a service, allowing stops
not served in the current day to still be output.
There is also a new parameter allowing to output stops even if there is only
arrival edges (and no departure edges) linked to it.