Feature : new parameters on StopAreasListFunction.
We can ask to hide stop areas that are not served by a service (departure
and arrival forbidden or no service in the next X days (configurable too)).
We can also allow output of lines only arriving on the stop area, i.e. on
the terminus of a one way line.