Bugfix in TridentFileFormat.
Some stops were in the output file as a Trident StopPoint but not as a
Trident StopArea (linestops with arrival and departure forbidden).
Trident can handle this, know those stops are output as a StopArea and a
StopPoint, with a BoardingAlightPossibility tag inside VehicleJourneyAtStop.
Also add the possibility of importing files with BoardingAlightPossibility with 
the value "NeitherBoardOrAlight".