Bugfix on houses offset and coordinates calculation.
Handle in OSM datas houses in decreasing order, flip
start and end houses number in that case.
Correct coordinates by offset on a road chunk. SYNTHESE
was retrieving coordinates by the global offset on the road
instead of the relative offset on the chunk itself.
Improve offset on the edge of the chunk. If there is only
one house number on the chunk, place it on the middle of it.