Fixed admin site crash in the display of a route
This is a temporary work around that does not solve the underlying
issue. What happens is that we have a Path that is erased by the 
cleanup of and env but in a ContinuousService unlink we try to
remove ourself from the 'deleted' path. The crash happens in 
Path::removeService (the sharedServicesMutex is corrupt).

In this work around we load the services in the global env instead
of a local one and thus avoid the issue.