A temporary fix for alert messages.
This patch parses the output of the CMS to get back information
in the C++ side of RoutPlannerfunction.

The issue is that the information whether or not we got a message
is defined by the user in a CMS page but the RoutPlannerFunction
is supposed to return in DATA_MAX_WARNING_LEVEL_ON_LINE the fact
that there is a warning on a line. What we miss is a way from a
CMS page to change globally the top level parameters map.

** DO NOT MERGE ON TRUNK **