Bug fix for OSMFileFormat importer on house numbers.
In OSM datas houses are objects with coordinates, and they are linked to a
way in no particular order. Since we're only keeping maximum and minimum bounds
for each road chunk we need to check if the direction of the way is the same than
the direction of house numbers.
To ensure bounds are correct we loop over every road chunk and compare houses coordinates
to the start and end of the chunk and reverse bounds if necessary.