Bug fix on StopPoint positioning.
StopPoint weren't accessible from both end of the chunk they were projected on.
The metric_offset stored in database was also different from metric_offset of
houses in places_list. In order to fix this problem the metric_offset is now relative
to the entire Road, like it is for houses.