Feature: complete geometry of approach journey.
Before SYNTHESE couldn't extract the part of the geometry at the start of
an approach journey (from a house somewhere on a chunk to the first crossing)
or the end (from a crossing to the projected point of a stop point).
Now, with A* approach, it will build a custom geometry taking into account
those missing chunks.