merged from trunk r10358
Crash fix in geos distanceOp.cpp when compiling in releasewithdbginfo.

Same as https://extranet.rcsmobility.com/issues/12100 a dynamic_cast in
a static function was returning NULL instead of the correct object
even if it was one of those, causing an assert(0) to be called.

GEOS issue : http://trac.osgeo.org/geos/ticket/535