swapped 2 mutexes to avoid a dead lock
In case of a synthese instance being used in route search and
with inter synthese real time, the 2 mutextes are taken in an
order in route searching and in another in intersynthese update.

Not sure this is the only occurance of this problem as we have
never tested synthese with this workload.