kiosk, add multi-screen support
It is now possible to specify the screen where the browser must
be started. You set it in the "displays" configuration of you
kiosk_config.json like this:

...
"displays": {
    "DEF": {
      "url": "/onboard/unites-exploitation",
      "screen": "LVDS1"
    }
  }
...

To find the availaible screens on your system use the xrandr command.

This implementation is linux specific. It requires the commands 'xrandr'
and 'wmctrl' to be installed.