IneoFileFormat importer, add of a parameter to not merge stops
By default the importer does not create a new journey if the imported
stop are in the same stop area of the existing journey. This breaks
IneoBDSI as it won't reconciles.

This patch add a new parameter 'allow_different_stop_points_in_same_stop_area'
which is true by default to be compatible with the old behavior.

If this parameter is set to true then imported journey must match excactly the
stops to be reused.

Fixes bug ref #48531