Merged revision(s) 10860 from synthese3/branches/trunk_rcs/src: Merged revision(s) 10840-10859 from synthese3/branches/t2c2/src: IneoFileFormat importer, add of a parameter to not merge stops
By default the importer does not create a new journey if the imported
stop are in the same stop area of the existing journey. This breaks
IneoBDSI as it won't reconciles.

This patch add a new parameter 'allow_different_stop_points_in_same_stop_area'
which is true by default to be compatible with the old behavior.

If this parameter is set to true then imported journey must match excactly the
stops to be reused.

Fixes bug ref #48531

........
Feature : Refactoring of rollingStock to use boost::fusion for db

........
Feature : Refactoring of rollingStock to use boost::fusion for db

........
Merged revision(s) 10835 from synthese3/branches/trunk_rcs/src:
Merged revision(s) 10834 from synthese3/branches/t2c2/src:
Bugfix : Set links to all nodes, use getParent instead of getLink

Issue : #47363

........

........
........
merge commit r10840 from branches/t2c2

IneoFileFormat importer, add of a parameter to not merge stops

By default the importer does not create a new journey if the imported
stop are in the same stop area of the existing journey. This breaks
IneoBDSI as it won't reconciles.

This patch add a new parameter 'allow_different_stop_points_in_same_stop_area'
which is true by default to be compatible with the old behavior.

If this parameter is set to false then imported journey must match excactly the
stops to be reused.

Fixes bug ref #48531


........
Bugfix : export of the SentScenario
........
Bugfix : use of getLineStops instead of old getEdges
........

........