Message display verification
Modification of DisplayScreen::displaysMessage :
When a message display is associated to a display type, it is supposed to be displayed on all displayscreens which type is equal to the one associated to the current message.

Modification of CustomBroadcastPoint::displaysMessage :
When a message display is associated to all kind of display or all custom broadcast points, it is supposed to be displayed what ever the custom broadcast point is.

Issue : #46674