fix a bug in which a realtime vertex change is not accounted in a tdg
In getFromPresenceTime() we create a ServicePointer which is initialized with
RTData=False even is we have an assigned realTimeVertex.

This is due to the call of SchedulesBasedService::hasRealTimeData() which does
not check the realTimeVertex. This still needs to be fixed, there are propably
other bug in case of a real time vertex change.