Bugfix : Delete requests are not correctly processed
When the list of recipients of a delete request is a partial match of an existing message,
SYNTHESE removed those recipients and updated this message. However it also removed display
screen recipients, when prevented further delete requests to match the message.

Example :
1) An Ineo create message request generates a message with 2 recipients of type 'line'
(A and B) et 1 recipient of type 'displayscreen' (Type).
2) An Ineo delete message request is received for this message, with recipient A
=> SYNTHESE removed A and Type, leaving only B
3) An Ineo delete message request is received for this message, with recipient B
=> SYNTHESE searched a message with recipients B and Type and found no match

Fix : do not consider 'displayscreen' recipients in this algorithm