tc-235 additional fix
Adding timezone offset is not enough. Actually, Date.parse called with
an ISO-8601 string with unspecified TZ fallbacks to a default behaviour
which is different under linux (UTC) and windows (local time).

A portable way to do this is to convert manually TZ offset to iso8601 format to remove the ambiguity.