Bugfix tc-336 (report of #60403) :Infinite loop in route planning algorithm
A wrong boundary check caused the algorithm to loop on the same time slice forever in some situations. The boundary check has been fixed.
I also implemented a safe guard that prevents loops in the time loop of TimeSlotRoutePlanner : when the algorithm would go back in time,
the route planning request is judged "failed", the result set is cleared and an empty result is returned. An error log is printed.