Bugfix tc-325 & tc-326 : Proper implementation of the cache of currently activated messages
Now the cache relies on a simpler operator (UnicityAlarmLess) that prevents collision between messages but does not sort them.
This allows modifying and deleting of messages without corrupting the cache.
Messages deleted from the registry are properly removed from the cache.

The methods GetActivatedMessages and GetActivatedMessagesAt rely on another operator (ImportanceAlarmLess) that implements sorting
because the service 'messages' expects them to. Sorting is performed as it was before, except that commercial line comparison has
been removed because it was buggy and made no cense in many cases. In case of equality, messages are sorted by roids and pointers.

6 new tests have been added on the broadcasting. Many Robot keywords have been modified in an attempt to remove usage of global
variables and to allow testing with multiple messages.